requirements.txt
notebooklm-py[browser]>=0.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
aiofiles>=24.1.0
python-multipart>=0.0.9
pyinstaller>=6.8.0
